Create login controller with validation and tests
