+++ src/schemas/todos/shared/requestIdSchema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { idParamsSchema } from "../shared/idParamsSchema";
+import { idParamsSchema } from "./idParamsSchema";
 
 export const requestIdSchema = z.object({
   params: z.object({ id: idParamsSchema }),
